refactor: config
